Refactor - move entity interfaces/type
