+++ src/entities/Genre.ts
@@ -1,0 +1,6 @@
+
+export type Genre = {
+  id: number;
+  name: string;
+  image_background: string;
+};
