+++ src/entities/Platform.ts
@@ -1,0 +1,6 @@
+
+export type Platform = {
+  id: number;
+  name: string;
+  slug: string;
+};
