+++ src/hooks/useGames.ts
@@ -2,22 +2,11 @@ import {  useInfiniteQuery, useQuery } from "@tanstack/react-query";
 // import { GameQuery } from "../App";
 import APIclient, { FetchResponse } from "../services/api-client";
 
-import { Platform } from "./usePlatforms";
 import ms from "ms";
 import useGameQueryStore from "../store";
+import { Game } from "../entities/Game";
 
 const apiClient = new APIclient<Game>('/games')
-
-export type Game = {
-  id: number;
-  name: string;
-  slug: string;
-  description_raw: string;
-  background_image: string;
-  parent_platforms: { platform: Platform }[];
-  metacritic: number;
-  rating_top: number;
-};
 
 const useGames = () => 
   {
