+++ src/main.tsx
@@ -6,7 +6,6 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 
 
-import App from "./App.tsx";
 import theme from "./theme.ts";
 import "./index.css";
 import { RouterProvider } from "react-router-dom";
